requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
Jinja2==3.1.4
python-multipart==0.0.9
redis==5.0.7
SQLAlchemy==2.0.32
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
scikit-learn==1.5.1
joblib==1.4.2
pyyaml==6.0.2
scapy==2.5.0
